Using the service instead of ApplicationDbContext. Moved some files. Deleted some unused actions and views.
