+++ QuoteApp/Controllers/HomeController.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Mvc;
-using QuoteApp.Models;
+﻿using System.Web.Mvc;
+using QuoteApp.Database.Home;
 
 namespace QuoteApp.Controllers
 {
@@ -12,20 +8,6 @@ namespace QuoteApp.Controllers
         public ActionResult Index()
         {
             return View(new HomeViewModel());
-        }
-
-        public ActionResult About()
-        {
-            ViewBag.Message = "Your application description page.";
-
-            return View();
-        }
-
-        public ActionResult Contact()
-        {
-            ViewBag.Message = "Your contact page.";
-
-            return View();
         }
     }
 }
