+++ QuoteApp/QuoteApp.csproj
@@ -190,7 +190,6 @@
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
     <Compile Include="Helpers\PdfHelper.cs" />
-    <Compile Include="Helpers\Quarter.cs" />
     <Compile Include="Migrations\201412121540416_InitialCreate.cs" />
     <Compile Include="Migrations\201412121540416_InitialCreate.Designer.cs">
       <DependentUpon>201412121540416_InitialCreate.cs</DependentUpon>
@@ -288,9 +287,6 @@
       <DependentUpon>201501260757566_InvoiceDetailNotReq.cs</DependentUpon>
     </Compile>
     <Compile Include="Migrations\Configuration.cs" />
-    <Compile Include="Models\AccountViewModels.cs" />
-    <Compile Include="Models\HomeViewModel.cs" />
-    <Compile Include="Models\IdentityModels.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
   </ItemGroup>
@@ -388,8 +384,6 @@
     <Content Include="Views\_ViewStart.cshtml" />
     <Content Include="Views\Shared\Error.cshtml" />
     <Content Include="Views\Shared\_Layout.cshtml" />
-    <Content Include="Views\Home\About.cshtml" />
-    <Content Include="Views\Home\Contact.cshtml" />
     <Content Include="Views\Home\Index.cshtml" />
     <Content Include="Views\Account\_ChangePasswordPartial.cshtml" />
     <Content Include="Views\Account\_ExternalLoginsListPartial.cshtml" />
@@ -523,6 +517,12 @@
     <EmbeddedResource Include="Migrations\201501260757566_InvoiceDetailNotReq.resx">
       <DependentUpon>201501260757566_InvoiceDetailNotReq.cs</DependentUpon>
     </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\QuoteApp.Database\QuoteApp.Database.csproj">
+      <Project>{aba39d4d-dfde-4985-9334-266c8dbc8042}</Project>
+      <Name>QuoteApp.Database</Name>
+    </ProjectReference>
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
