+++ QuoteApp/Views/Home/Index.cshtml
@@ -2,7 +2,7 @@
     ViewBag.Title = "Home Page";
 }
 @using QuoteApp.Models
-@model QuoteApp.Models.HomeViewModel
+@model QuoteApp.Database.Home.HomeViewModel
 <div class="form-horizontal">
     <div class="form-group">
         <div class="col-lg-12">
